requirements.txt
fastapi
uvicorn
pandas
matplotlib
pydantic
python-dotenv
streamlit
sqlparse
tabulate
openai
